backend/requirements.txt
Django==4.1
mysqlclient==2.1.1
django-environ==0.9.0
djangorestframework==3.13.1
django-debug-toolbar